openvibe :  * modified cmake file configurations  + added per dependancy module .cmake file
openvibe-kernel :
 * enhanced windows portability
 * modified cmake file configurations
 + added per dependancy module .cmake file

openvibe-toolkit :
 * enhanced windows portability
 * modified cmake file configurations
 + added per dependancy module .cmake file

openvibe-modules-ebml :
 * enhanced windows portability
 * modified cmake file configurations

openvibe-modules-fs :
 * modified cmake file configurations

openvibe-modules-socket :
 * modified cmake file configurations

openvibe-modules-system :
 * modified cmake file configurations

openvibe-plugins-acquisition :
 * modified cmake file configurations
 + added per dependancy module .cmake file

openvibe-plugins-file-io :
 * modified cmake file configurations
 + added per dependancy module .cmake file

openvibe-plugins-samples :
 * modified cmake file configurations
 + added per dependancy module .cmake file

openvibe-plugins-signal-processing :
 * modified cmake file configurations
 + added per dependancy module .cmake file

openvibe-plugins-simple-visualisation :
 * modified cmake file configurations
 + added per dependancy module .cmake file

openvibe-plugins-tools :
 * modified cmake file configurations
 + added per dependancy module .cmake file

openvibe-applications-acquisition-server :
 * modified cmake file configurations
 + added per dependancy module .cmake file

openvibe-applications-designer :
 * modified cmake file configurations
 + added per dependancy module .cmake file


git-svn-id: svn://scm.gforge.inria.fr/svn/openvibe@298 c330d7e9-fc0c-0410-a5b3-fd85c6f5aa8f

+++ openvibe-toolkit/trunc/src/openvibe-toolkit/reader/ovtkIBoxAlgorithmSignalInputReaderCallback.cpp
@@ -114,7 +114,7 @@ void CBoxAlgorithmSignalInputReaderCallback::processChildData(const void* pBuffe
 	else
 	{
 		// ...
-		if(l_rTop==OVTK_NodeId_Signal_SamplingRate) m_rCallback.setSamplingRate(m_pReaderHelper->getUIntegerFromChildData(pBuffer, ui64BufferSize));
+		if(l_rTop==OVTK_NodeId_Signal_SamplingRate) m_rCallback.setSamplingRate(static_cast<uint32>(m_pReaderHelper->getUIntegerFromChildData(pBuffer, ui64BufferSize)));
 	}
 }
